Reorganize source code, create DataStream directory
