+++ DataStream/Compression.cc
