+++ DataStream/Smoothing.cc
