+++ DataStream/Smoothing.h
